requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
faker
numpy
python-dateutil
